13980425 رفع ایردات مربوط به ثبت سفارش و سبد خرید و ...

+++ app/src/main/java/ir/rayas/app/citywareclient/Adapter/RecyclerView/BusinessNoCommissionReceivedRecyclerViewAdapter.java
@@ -11,6 +11,10 @@ import android.widget.RadioButton;
 import android.widget.RelativeLayout;
 
 
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
+
+import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -21,6 +25,7 @@ import ir.rayas.app.citywareclient.Share.Layout.View.TextViewPersian;
 import ir.rayas.app.citywareclient.Share.Utility.Utility;
 import ir.rayas.app.citywareclient.View.MasterChildren.ShowBusinessCommissionActivity;
 import ir.rayas.app.citywareclient.View.MarketerChildren.FactorDetailsActivity;
+import ir.rayas.app.citywareclient.ViewModel.Marketing.BusinessCommissionAndDiscountViewModel;
 import ir.rayas.app.citywareclient.ViewModel.Marketing.MarketingPayedBusinessViewModel;
 
 
@@ -74,7 +79,6 @@ public class BusinessNoCommissionReceivedRecyclerViewAdapter extends RecyclerVie
     @Override
     public void onBindViewHolder(final ViewHolder holder, final int position) {
 
-
         holder.FullNameTextView.setText(ViewModelList.get(position).getCustomerFullName());
         holder.TicketNumberTextView.setText(ViewModelList.get(position).getTicket());
 
@@ -120,7 +124,7 @@ public class BusinessNoCommissionReceivedRecyclerViewAdapter extends RecyclerVie
                 }  else {
                     ViewModel.get(position).setSelected(true);
                 }
-                Context.SetViewPriceInFooter(ViewModelList.get(position).getPrice(),ViewModel.get(position).getSelected());
+                Context.SetViewPriceInFooter((int)ViewModelList.get(position).getPrice(),ViewModel.get(position).getSelected());
                 holder.BusinessCommissionSelectedRadioButton.setChecked(ViewModel.get(position).getSelected());
             }
         });
